Move module exports and imports to surface
