+++ crates/analyzer/src/surface.rs
@@ -6,6 +6,7 @@ mod trees;
 
 use std::sync::Arc;
 
+use files::FileId;
 use lower::SurfaceContext;
 pub use source_map::*;
 pub use trees::*;
@@ -18,6 +19,12 @@ use crate::{
 
 #[salsa::query_group(SurfaceStorage)]
 pub trait SurfaceDatabase: SourceDatabase + ResolverDatabase {
+    #[salsa::invoke(ModuleImports::module_imports_query)]
+    fn module_imports(&self, file_id: FileId) -> Arc<ModuleImports>;
+
+    #[salsa::invoke(ModuleExports::module_exports_query)]
+    fn module_exports(&self, file_id: FileId) -> Arc<ModuleExports>;
+
     #[salsa::invoke(SurfaceContext::data_surface_query)]
     fn data_surface(&self, id: InFile<DataGroupId>) -> Arc<WithArena<DataGroup>>;
